+++ src/models/user.rs
@@ -2,13 +2,13 @@ use serde::{Serialize, Deserialize};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct User {
-    pub id: i32,
+    pub id: i64,
     pub full_name: String,
     pub profile_pic: String,
     pub cpf: String,
     pub birthdate: String, // Use a string in ISO 8601 format
     pub marital_status: String,
-    pub gross_mensal_income: i32,
+    pub gross_mensal_income: i64,
     pub email: String,
     pub phone_number: String,
     pub is_admin: bool,
